hello-mcp-python/pyproject.toml
[project]
name = "hello-mcp-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.6.0",
    "pillow>=11.2.1",
]
